changed unsigned int and double data types to int, might be changed later

+++ include/ecs/PlayerStats.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 struct PlayerStats{
-    unsigned int maxBombas = 3;
-    unsigned int rangoExplosion = 3;
-    double velocidad = 3;
+    int maxBombas = 3;
+    int rangoExplosion = 3;
+    int velocidad = 3;
 };
 
 struct PlayerConfig{
